Introduce initial transition, and refactor a little.
