+++ app/src/main/java/com/hedvig/app/feature/embark/EmbarkModel.kt
@@ -1,0 +1,10 @@
+package com.hedvig.app.feature.embark
+
+import com.hedvig.android.owldroid.graphql.EmbarkStoryQuery
+import com.hedvig.app.util.Percent
+
+data class EmbarkModel(
+    val passage: EmbarkStoryQuery.Passage?,
+    val navigationDirection: NavigationDirection,
+    val progress: Percent,
+)
