+++ app/src/main/java/com/hedvig/app/feature/embark/ExpressionResult.kt
@@ -1,0 +1,9 @@
+package com.hedvig.app.feature.embark
+
+sealed class ExpressionResult {
+    data class True(
+        val resultValue: String?,
+    ) : ExpressionResult()
+
+    object False : ExpressionResult()
+}
