+++ app/src/main/java/com/hedvig/app/feature/embark/NavigationDirection.kt
@@ -1,0 +1,7 @@
+package com.hedvig.app.feature.embark
+
+enum class NavigationDirection {
+    INITIAL,
+    FORWARDS,
+    BACKWARDS,
+}
